Only show pay button if someone needs payment
